# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Collection Service Documentation",
        "version": "1.0.0"
    },
    "paths": {
...
        "/transfer/create-recipient": {
            "post": {
                "tags": [
                    "Transfer"
                ],
                "summary": "",
                "description": "",
                "operationId": "createTransferRecipient",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "name",
                                    "accountNumber",
                                    "bankCode"
                                ],
                                "properties": {
                                    "name": {
                                        "description": "The recipient's name according to their account registration",
                                        "type": "string",
                                        "example": "Tolu Robert"
                                    },
                                    "accountNumber": {
                                        "description": "",
                                        "type": "string",
                                        "example": "0000000000"
                                    },
                                    "bankCode": {
                                        "description": "",
                                        "type": "string",
                                        "example": "057"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success, Failure, or Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/FailedResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ValidationErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
        "/transfer/initiate-via-recipient": {
            "post": {
                "tags": [
                    "Transfer"
                ],
                "summary": "Initiate a bank transfer via recipient code",
                "description": "Send money to your customers.",
                "operationId": "initiateTransferViaRecipientCode",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "required": [
                                    "recipientCode",
                                    "amountInKobo"
                                ],
                                "properties": {
                                    "recipientCode": {
                                        "description": "Amount in kobo",
                                        "type": "string",
                                        "example": "1000"
                                    },
                                    "amountInKobo": {
                                        "description": "Amount in kobo",
                                        "type": "string",
                                        "example": "1000"
                                    },
                                    "description": {
                                        "description": "Reason from the transfer",
                                        "type": "string",
                                        "example": "payment for order 1234"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Success, Failure, or Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "oneOf": [
                                        {
                                            "$ref": "#/components/schemas/SuccessResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/FailedResponse"
                                        },
                                        {
                                            "$ref": "#/components/schemas/ValidationErrorResponse"
                                        }
                                    ]
                                }
                            }
                        }
                    },
                    "401": {
                        "description": "Unauthorized - Invalid authentication",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/UnauthorizedResponse"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "BearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "SuccessResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "200"
                    },
                    "message": {
                        "type": "string",
                        "example": "Request processed successfully"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "FailedResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "300"
                    },
                    "message": {
                        "type": "string",
                        "example": "Failed to process request"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "ValidationErrorResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "payloadValidationError"
                    },
                    "message": {
                        "type": "string",
                        "example": "Validation error"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            },
            "UnauthorizedResponse": {
                "properties": {
                    "status": {
                        "type": "string",
                        "example": "401"
                    },
                    "message": {
                        "type": "string",
                        "example": "Unauthorized access"
                    },
                    "data": {
                        "type": "object"
                    }
                },
                "type": "object"
            }
        },
        "securitySchemes": {
            "BearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}